Migrations to enable "other organisations"
Have added a string field "other_organisation" to users. This is
intended to contain the name of the user's organisation when that
organisation is not listed in the organisations table. A boolean field
"is_other" has been added to the organisations table, to indicate any
dummy organisations whose selection indicates a need to use the
"other_organisation" field.