Plan status view
- bundle install
- rake db:migrate
Creates a view summarising the status of a plan (when sections were
locked and by whom, when questions were answered and by whom). Uses the
Schema Plusgem to generate the views (two supporting views needed) via
migrations. The idea is to improve the performance of the plan editing
page by reducing server load.
Also includes the Better Errors gem.