Export settings fixed for modal
Export settings were working well on Export tab, but not in modal on
the edit tab. Fixed this by having form submit via AJAX, using JS to
reset fields to default values when “Reset” is clicked, collapsing
accordion on save, other JS to tweak other on-screen notifications/cues
as appropriate.