Values used in space calculation as constants
Intended to allow easier tweaking of calculation. Actual logic of
calculation may need to be adjusted to account for (variation in) line
spacing for different paragraph styles.

Note: number of pages allowed for each template should be recorded in
the settings table as follows (as per reference in plan.rb to
self.dmptemplate.settings(:export).max_pages):

export:
  :max_pages: 3