Fixed flash not showing for reCAPTCHA failures.
If the user does not enter a valid reCAPTCHA then no message is
displayed. The problem seemed to be with
app/controllers/contacts_controller.rb which sets flash values
with the key :error. According to the Rails documentation on
the [flash](http://guides.rubyonrails.org/action_controller_overview.html#the-flash),
only :notice and :alert are supported:

> Note that it is also possible to assign a flash message as part
> of the redirection. You can assign :notice, :alert or the
> general purpose :flash:

Replaced flash[:error] with flash[:alert] so the flashes appear
on the title screen.